+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.22.0] - 2022-01-03
 ### Added
-- Loading page
+- Loading page.
+- Polling for current electricity widgets.
 
 ### Fixed
 - Portal logged out page.
+- Refreshing widgets on Dashboard.
 
 ## [1.21.1] - 2022-01-01
 ### Fixed